Learn by Doing v1/backend/requirements.txt
# Andromeda SPED App Backend Requirements
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
httpx>=0.28.0
pytest>=8.3.0
pytest-cov>=6.0.0
black>=24.10.0
flake8>=7.1.0
mypy>=1.13.0
sendgrid>=6.11.0
boto3>=1.28.0
cryptography>=42.0.0
